split test_gpt_oss_4gpu into bf16 + mxfp4
