+++ test/registered/models_e2e/test_gpt_oss_4gpu_mxfp4.py
@@ -3,21 +3,11 @@ import unittest
 from sglang.test.ci.ci_register import register_cuda_ci
 from sglang.test.gpt_oss_common import BaseTestGptOss
 
-register_cuda_ci(est_time=392, stage="base-c", runner_config="4-gpu-h100")
-register_cuda_ci(est_time=350, stage="base-c", runner_config="4-gpu-b200")
+register_cuda_ci(est_time=220, stage="base-c", runner_config="4-gpu-h100")
+register_cuda_ci(est_time=220, stage="base-c", runner_config="4-gpu-b200")
 
 
-class TestGptOss4Gpu(BaseTestGptOss):
-    def test_bf16_120b(self):
-        self.run_test(
-            model_variant="120b",
-            quantization="bf16",
-            expected_score_of_reasoning_effort={
-                "low": 0.58,
-            },
-            other_args=["--tp", "4", "--cuda-graph-max-bs", "200"],
-        )
-
+class TestGptOss4GpuMxfp4(BaseTestGptOss):
     def test_mxfp4_120b(self):
         self.run_test(
             model_variant="120b",
